Certifier user header refactoring
Refactored Certifier headers. All Certifier user facing public
interfaces are moved to certifier_framework.h and all the Certifier
utilities are moved to certifier_utilities.h. No other headers should be
used by user applications. As a first step, some internal headers now
include the public headers to maintain consistency. These will be
cleaned up in the followup patches.
