+++ include/support.h
@@ -13,8 +13,6 @@
 #include <cmath>
 #include <string>
 
-#include "certifier.pb.h"
-
 #ifndef byte
 typedef unsigned char byte;
 #endif
@@ -23,13 +21,15 @@ typedef unsigned char byte;
 #include <openssl/rsa.h>
 #include <openssl/x509.h>
 #include <openssl/evp.h>
-#include "certifier.pb.h" 
-#include "certifier.h" 
+#include "certifier.pb.h"
+#include "certifier.h"
 
 #ifdef OE_CERTIFIER
 #include "openenclave/attestation.h"
 #include "openenclave/sealing.h"
 #endif
+
+#include "certifier_utilities.h"
 
 using std::string;
 
@@ -50,27 +50,11 @@ using std::string;
 
 #ifndef _SUPPORT_H__
 #define _SUPPORT_H__
-const int block_size = 16;
-const int num_bits_in_byte = 8;
-
-bool write_file(const string& file_name, int size, byte* data);
-int file_size(const string& file_name);
-bool read_file(const string& file_name, int* size, byte* data);
-bool read_file_into_string(const string& file_name, string* out);
 
 bool encrypt(byte* in, int in_len, byte *key,
             byte *iv, byte *out, int* out_size);
 bool decrypt(byte *in, int in_len, byte *key,
             byte *iv, byte *out, int* size_out);
-
-bool digest_message(const char* alg, const byte* message, int message_len,
-    byte* digest, unsigned int digest_len);
-
-
-bool authenticated_encrypt(const char* alg, byte* in, int in_len, byte *key,
-            byte *iv, byte *out, int* out_size);
-bool authenticated_decrypt(const char* alg, byte* in, int in_len, byte *key,
-            byte *out, int* out_size);
 
 bool make_certifier_rsa_key(int n,  key_message* k);
 bool rsa_public_encrypt(RSA* key, byte* data, int data_len, byte *encrypted, int* size_out);
@@ -102,13 +86,6 @@ bool generate_new_rsa_key(int num_bits, RSA* r);
 bool key_to_RSA(const key_message& k, RSA* r);
 bool RSA_to_key(const RSA* r, key_message* k);
 
-EC_KEY* generate_new_ecc_key(int num_bits);
-EC_KEY* key_to_ECC(const key_message& kr);
-bool ECC_to_key(const EC_KEY* e, key_message* k);
-
-bool private_key_to_public_key(const key_message& in, key_message* out);
-bool get_random(int num_bits, byte* out);
-
 bool make_key_entity(const key_message& key, entity_message* ent);
 bool make_measurement_entity(const string& measurement, entity_message* ent);
 bool make_property(string& name, string& type, string& cmp, uint64_t int_value,
@@ -128,7 +105,6 @@ bool make_indirect_vse_clause(const entity_message& subject, string& verb,
 bool make_claim(int size, byte* serialized_claim, string& format, string& descriptor,
     string& not_before, string& not_after, claim_message* out);
 bool get_vse_clause_from_signed_claim(const signed_claim_message& scm, vse_clause* c);
-void print_bytes(int n, byte* buf);
 void print_key_descriptor(const key_message& k);
 void print_entity_descriptor(const entity_message& e);
 void print_property_descriptor(const property& p);
@@ -146,46 +122,10 @@ bool make_signed_claim(const char* alg, const claim_message& claim, const key_me
 bool verify_signed_claim(const signed_claim_message& claim, const key_message& key);
 bool get_vse_clause_from_signed_claim(const signed_claim_message& scm, vse_clause* c);
 
-// Serialized time: YYYY-MM-DDTHH:mm:ss. sssZ
-bool time_now(time_point* t);
-bool time_to_string(time_point& t, string* s);
-bool string_to_time(const string& s, time_point* t);
-bool add_interval_to_time_point(time_point& t_in, double hours, time_point* out);
-int compare_time(time_point& t1, time_point& t2);
-void print_time_point(time_point& t);
-void print_entity(const entity_message& em);
-void print_key(const key_message& k);
-void print_rsa_key(const rsa_message& rsa);
-void print_ecc_key(const ecc_message& rsa);
-void print_platform(const platform& pl);
-void print_environment(const environment& env);
-void print_property(const property& prop);
-
-// X509 artifact
-bool produce_artifact(key_message& signing_key, string& issuer_name_str,
-      string& issuer_description_str, key_message& subject_key,
-      string& subject_name_str, string& subject_description_str,
-      uint64_t sn, double secs_duration, X509* x509, bool is_root);
-bool verify_artifact(X509& cert, key_message& verify_key,
-    string* issuer_name_str, string* issuer_description_str,
-    key_message* subject_key, string* subject_name_str, string* subject_description_str,
-    uint64_t* sn);
-
-int cipher_block_byte_size(const char* alg_name);
-int cipher_key_byte_size(const char* alg_name);
-int digest_output_byte_size(const char* alg_name);
-int mac_output_byte_size(const char* alg_name);
-
-bool asn1_to_x509(const string& in, X509 *x);
-bool x509_to_asn1(X509 *x, string* out);
-bool make_root_key_with_cert(string& type, string& name, string& issuer_name, key_message* k);
-
 int sized_pipe_read(int fd, string* out);
 int sized_pipe_write(int fd, int size, byte* buf);
 int sized_ssl_read(SSL* ssl, string* out);
 int sized_ssl_write(SSL* ssl, int size, byte* buf);
-int sized_socket_read(int fd, string* out);
-int sized_socket_write(int fd, int size, byte* buf);
 
 class cert_keys_seen {
 public:
