+++ sample_apps/analytics_example/enclave/ecalls.cc
@@ -7,9 +7,6 @@
 #include <openenclave/attestation/verifier.h>
 #include <openenclave/bits/report.h>
 #include <openenclave/bits/module.h>
-#include "certifier.h"
-#include "support.h"
-#include "simulated_enclave.h"
 
 #include <sys/socket.h>
 #include <arpa/inet.h>
@@ -23,10 +20,8 @@
 #include <openssl/hmac.h>
 #include <openssl/err.h>
 
-#include "certifier.h"
-#include "support.h"
-#include "simulated_enclave.h"
-#include "cc_helpers.h"
+#include "certifier_framework.h"
+#include "certifier_utilities.h"
 #include "../policy_key.cc"
 
 #include "analytics_app.cc"
