+++ sample_apps/simple_app/example_app.cc
@@ -27,10 +27,7 @@
 #include <openssl/hmac.h>
 #include <openssl/err.h>
 
-#include "support.h"
-#include "certifier.h"
-#include "simulated_enclave.h"
-#include "cc_helpers.h"
+#include "certifier_framework.h"
 
 
 // operations are: cold-init, warm-restart, get-certifier, run-app-as-client, run-app-as-server
