+++ sample_apps/simple_app_under_app_service/service_example_app.cc
@@ -27,11 +27,8 @@
 #include <openssl/hmac.h>
 #include <openssl/err.h>
 
-#include "support.h"
-#include "certifier.h"
-#include "simulated_enclave.h"
-#include "application_enclave.h"
-#include "cc_helpers.h"
+#include "certifier_framework.h"
+#include "certifier_utilities.h"
 
 // operations are: cold-init, warm-restart, get-certifier, run-app-as-client, run-app-as-server
 DEFINE_bool(print_all, false,  "verbose");
